Make alias generation generic
